+++ src/main/java/com/nhnacademy/illuwa/d_review/entity/review/Review.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.illuwa.review.entity;
 
-import com.nhnacademy.illuwa.book.Book;
+import com.nhnacademy.illuwa.book.entity.Book;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
